refactor: update funding API imports to new endpoint structure and add new funding endpoint files
