+++ packages/interface/src/api/funding/endpoint/apiFnd004.ts
