+++ packages/web/src/features/manage-club/funding/services/useGetNewFundingList.ts
@@ -1,7 +1,7 @@
 import apiFnd005, {
   ApiFnd005RequestQuery,
   ApiFnd005ResponseOk,
-} from "@sparcs-clubs/interface/api/funding/apiFnd005";
+} from "@sparcs-clubs/interface/api/funding/endpoint/apiFnd005";
 import { useQuery } from "@tanstack/react-query";
 
 import { mockupManageFunding } from "@sparcs-clubs/web/features/manage-club/services/_mock/mockManageClub";
