+++ packages/web/src/features/manage-club/funding/services/usePostFunding.ts
@@ -1,7 +1,7 @@
 import apiFnd001, {
   ApiFnd001RequestBody,
   ApiFnd001ResponseCreated,
-} from "@sparcs-clubs/interface/api/funding/apiFnd001";
+} from "@sparcs-clubs/interface/api/funding/endpoint/apiFnd001";
 import { useMutation } from "@tanstack/react-query";
 
 import {
