+++ packages/web/src/features/manage-club/funding/services/usePutFunding.ts
@@ -1,4 +1,4 @@
-import apiFnd003 from "@sparcs-clubs/interface/api/funding/apiFnd003";
+import apiFnd003 from "@sparcs-clubs/interface/api/funding/endpoint/apiFnd003";
 import { useMutation } from "@tanstack/react-query";
 import { z } from "zod";
